Update confidence threshold values and sync returned answer count with threshold answer count
Currently ConversationApiImpl is configured to return a certain number
of answers, depending on the confidence category.  And
CumulativeConfidenceCategorizer is configured to look at a certain
number of answers when determining the confidence category.  These two
numbers should be the same.

In order to avoid having to keep two configured numbers in sync, handle
trimming the answer list in the categorizer, rather than in the api
implementation.